moved fieldset, placeholder and input to subcomponent.js file, removed vendor prefixes, minor fixes
