--- components/bl-input-with-mask/src/placeholder.js
@@ -1,17 +1,0 @@
-const { cn } = BackendlessUI.CSSUtils;
-
-export const Placeholder = ({ placeholder, htmlFor, value, isFocused }) => {
-
-  return (
-    <label
-      htmlFor={ htmlFor }
-      className={ cn(
-        'input-with-mask__label',
-        {
-          'input-with-mask__label--move-up': isFocused || value,
-          'input-with-mask__label--focused': isFocused,
-        }) }>
-      { placeholder }
-    </label>
-  );
-};
